Do not unpack bundle each node deployment
With the previous changes to the renamed UnmanagedDeployer we were unpacking the
bundle for each node deployment. This is fine for production scenarios where we
will only be deploying a single node at a time but it creates unnecessary
overhead for embedded deployments. The unpacking of the bundle and the clean up
have been refactored out into separate methods so we can avoid that overhead.