store the hostname, thift, and native ports for each seed node
When the auto installer ran we updated rhq-server.properties with the hostnames
and thrift ports for each node, but we were not storing the native transport
(i.e., cql) port for each seed node.

I have refactored the logic to store both ports for each node. The format of
the rhq.cassandra.seeds property now looks like,

rhq.cassandra.seeds=127.0.0.1|9160|9042,127.0.0.2|9160|9042

Seed entries are comma-delimited and within each entry there are three
pipe-delimited fields. The first is the hostname, the second is the thrift port
and the third is the native transport port.

CassandraNode has been updated to store the native transport port and a method
has been added to parse the string into a CassandraNode object.