fixing logic for passing nodes to SchemaManager
A new constructor has been added to SchemaManager that takes a
List<CassandraNode>. CCMTestNGListener now just passes the list of nodes it
gets back from createCluster() into that constructor.

Also changing the logic that throws an exception if we think a cluster is
already running. We only want to throw the exception if the
-DskipClusterShutdown property is not defined.