only need to deploy cluster once for test suite
When we try to deploy the cluster a second time, something is going wrong and
we fail to install the schema the second time around. As of right now there is
no need to incur the overhead of deploying a new cluster for each class that
needs to talk to cassandra. We can revisit when the need arises, like for
failure and consistency testing.