fixing jsp compile errors
Some errors were occurring due to project dependencies with a scope of provided
not being added to the classpath for jspc. I compared the jenkins logs from
master and the cassandra-backend branch and I see in master where things
like core-domain, which has a provided scope, are getting added to the
classpath. Setting useProvidedScope to true in the jspc plugin configuration
does seem to get the classpath squared away.