clean up bundle recipe so it no longer creates incorrect directories
In the post-install target the recipe was creating some data and log
directories, and due to some prior changes those directories wound up with the
wrong paths. The paths are correct in cassandra.yaml though. That part of the
bundle recipe has been altgother removed though as cassandra will create the
directories.