[Bug 907407 - RHQ master build fails starting - Failed to load module: deployment.rhq.ear.rhq-enterprise-server-ejb3.jar:main] Server JAR has a dependency on the OJDBC driver because Quartz sometimes needs to load a BLOB class from the driver. Postgres installs will not have (and don't need) the Oracle driver, so we declared the dependency "optional" in jboss-deployment-structure.xml.  But "optional" just means AS7 will ignore the fact that the module may not exist at all.  Our build and install  process is simplified by always creating the oracle driver module structure. But "optional" will not ignore the fact that the module directory is empty (no jar file). So, this commit ensures an empty/mock OJDBC driver jar will be placed in the module directory if the driver is not available at build time.
The JAR has a readme file inside describing the process of how to install
a real driver (by providing a wiki link).  We felt it was easier to provide
the whole structure and just let users drop in the real driver as opposed to
not providing the module structure at all, and making them build it.  For users
following the Oracle setup/prerequiste docs the steps are unchanged. For postgres
installs these changes should just make things work.