initial commit for metrics-simulator module
metrics-simulator will produce a program that stores and aggregates metrics. It
provides a convenient, easy way to simulate load as well as way to observe data
moving through the various metrics tables without having to run the program for
extended periods of time. This will enable you to more easily observe performance
and what not in the face of things like compaction, flushing of memtables, data
distribution and so on.