disabling tests
So far I have been running the simulator via maven test execution for
expediency. The longer term intention those is to produce a stand-alone program
that can be run and configured with a json file that specifies the simulation.

The simulator is *not* intended to be run as part of the test suite.