add error handling to terminate simulation
Since smal heap sizes are often used, it is common for a node to go down with
an OutOfMemoryError. This commit adds error handling to terminate the simulation
when an error occurs. Right now there is an assumption that the whole cluster
stays up for the duration of the simulation in order for it to be considered a
good run.