Fix classloading issues when defining objects in the std script context.
It seems that the change of the container to AS7 has caused a regression
in the CLI alert scripts that were no longer able to properly define the
standard context provided to scripts by RHQ.

This is fixed by using a differently configured Javassist's ClassPool
during the various class modifications we use to prepare the *Manager
objects and when generating a resource proxy in ProxyFactory.

Also, the errors that can happen during the various stages of this setup
are now thrown up the chain instead of merely logged so that errors like
this are apparent - it is not something the user should be expected to
deal with.