[BZ 928971] - ClassPoolFactory is now resilient against context classloader changes
The instance must not be cached because it could be using a wrong context
classloader if that changed between invocations.

Also made changes to lower the numbder of instances of the ClassPool
in the created while using the ConfigurationClassBuilder.

This, in addition to commit fe8952cf9631c2d839f9e37d53f8c1aa1bb04b83
(which I failed to annotate with the BZ number), should fix the bug.