remove dependency on bundles for cassandra deployer code
UnmanagedDeployer has been renamed to Deployer. There is no longer notion of
managed deployment since Cassandra will not be installed via agent provisioning.
The deployment code has been greatly simplified by the removal of the bundle
deployment code. It was only being used for variable substitution. That is now
performed by TokenReplacingReader.

This is also the initial commit for DeployerITest which is an integration test
that verifies deployment is performed as expected.

Lastly, TokenRelacingReader has been updated to properly handle escaping
strings.