initial commit for server-control module which will replace rhq.sh uber script
The server-control maven module is deployed as a JBoss module and is executed
via the rhqctl.sh script. It is a lightweight wrapper around the scripts used
to install/start/stop RHQ services. It provides platform independent support
for several non-trivial things like parsing options and checking to see that
the server is initialized.