make stack size passed to C* JVM configurable
Cassandra sets the JVM stack size to 180k in cassandra-env.sh which causes a
segfault on OpenJDK 6. This commit parameterizes the value passed to the -Xss
option in cassandra-env.sh and also makes it configurable in the metrics
simulator. There is now a stackSize property similar to the heapSize property.
It defaults to 180k.

This commit also updates the simulator to shut down the cluster when the
simulation run finishes.