Bug 949082 RFE: Allow alert ack/delete from recent alerts portlet and reporting view
Based on contribution from mark Addy (thanks!)
Change the "Delete" and "Acknowledge" alert buttons in the AlertHistoryView
to use "ResourceAuthorizedTableAction" rather than  "AbstractTableAction".
This generates additional calls to the backend to check permissions on
row selections (for non-admins). But, it should likely not be noticeable.

I made one small optimization in ResourceAuthorizedTableAction.  If
constructed with a requiredPermission=null it will ignore the authz check.
Using that we can make it perform like a standard table action. So, if in
fact the user is an admin or otherwise has write access then it should
act as before.  I took away the isEnabled() overrides in the original
patch in favor of this approach.  The overrides had a small issue because
they did not consider "enablement". Meaning the buttons would be active
(for admins) even if nothing was selected.