initial commit for console command
This initial commit introduces the console command but it is not fully
functional yet. Also refactoring the stop command some so that it first checks
for a pid file before trying to stop a process. Also made some changes in the
stop command to get it do the right thing, which is stop whatever is installed
by default. There is still a bit of work to get that logic right.