Bug 872731
Using the new feature allowing timeouts on agent client calls, limit
individual live data queries to the agent to 10s, making this more
likely to deal with a slow agent.

Also, fix up the MeasurementDataManager Local and Remote. Deprecated the
live data remote methods, which should not be exposed due to the impact
on agents. Fixed some jdoc for the remote.