package and deploy datastax cassandra driver as jboss module
We had been packaging the driver inside the ear's lib directory. I am not sure
if it was happening before but we ran into a problem like the one described in
https://issues.jboss.org/browse/IMMUTANT-144. The issue involves multiple
deployments of netty and netty supplying a jboss-beans.xml. The most robust
solution is to deploy the driver as a module. Now both the RHQ installer and
EAR depend on the module which is only loaded once, eliminating the netty
problem. The JBoss AS deployment for server itests has also been updated so that
the driver module is deployed for itests.

For reasons yet to be determined, these problems have not manifested themselves
in jenkins builds.

The CassandraClusterManager.startCluster method has been updated to wait for
cluster initialization before returning in order to give Cassandra a chance to
complete the superuser schema initialization.

Lastly, the dev profile in the server-metrics module has been temporarily
commented out because the data-migrator self-executing jar was getting deployed
in the EAR. This was causing the EAR deployment to fail after I fixed the netty
problem. The two problems were unrelated, and we couldn't see the data
migrator problem because the deployment failed earalier on in the process.