check that the service is isntalled before executing console command
There are a couple other minor changes in this commit. I have updated the
container build script to set the autoinstall.enabled property to true. I don't
think it make sense for it to be false now that this is the only way of
installing the server. And I am removing the the cluster initialization check
from CCMTestNGLisetener to see if it fixes the error on jenkins.