Bug 906385 [RFE] Trait change alert conditions should support value expression
Trait Change alert conditions now optionally support a regex matching test
for the new value.  If omitted the alert condition is unchanged from prior
behavior, the alert condition matches if the trait value changes.

If provided, the alert condition matches if the trait value changes and if
the new value matches the regex pattern.

Existing alert definitions are unaffected.

Also:
- a little cleanup of relevant I18N messages