Bug 593932 - (PRODMGT-355) Allow Trap OID for SNMP notification to be overridden per-alert definition
Snmp trap OID is set on v2c and v3 traps as the second variable binding.

In the past, users had been confused by the name of the SNMP notification parameter 'oid'. This parameter actually defines a variable binding prefix for the RHQ alert bindings.

Added tests for this server plugin
Added documentation to the plugin descriptor