Bug 955816 - Bundle deployment to EAP 6 standalone server fails due to deployDir trait not being available
This fix adds a new pluginConfiguration property "deploymentScannerPath" which represents a user configurable location for where a deployment scanner will be monitoring. By default, a value is read from the deployment-scanner module if defined in the host/standalone.xml file.

If deployment-scanner can't be read or is undefined, "deploymentScannerPath" will default to null (unset).