use separate executor and stream handler when running rhq-installer.sh --test
In the startRHQServerForInstallation method we first start the server then
execute rhq-installer.sh --test to verify that the server is ready for the
installer to run. On multiple occassions I got a IllegalThreadStateException
due to trying to run threads in PumpStreamHandler that had already been
stopped. This commit runs rhq-installer.sh --test with a separate executor and
PumpStreamHandler to prevent this from happening.