enable thrift for tests and remove unnecessary column in metrics_index table
We want thrift turned on during development so we can use cqlsh and
cassandra-cli. There is no longer a need to the null_col colum in metrics_index
as tables can now consist solely of columns that are part of the PK. This will
reduce space on disk as well as memory.