updating schema for metrics_index table so that rows are not reused
The way in which the metrics_index table has been used followed a big C*
anti-pattern. There was was one row devoted for each bucket that has data to be
aggregated. Once all of the schedules in a given row had their data aggregated
all of the columns in the row were deleted. This is increasingly inefficient as
the data set grows for a couple reasons. First, reads become slower because C*
has to iterate over more and more tombstone columns. Secondly, those tombstone
columns use up memory just like live columns do.

The columns stored are the same, but we are now using a compound row/partition
key that consists of the bucket name and a timestamp. The timestamp is the
start of an interval for a given bucket. So let's say we store some raw data at
12:22 pm. We will then update the {'one_hour_metrics', 12:00} row in the
metrics_index table for that schedule id. After 13:00, that will will never be
written to again and when aggregation runs over that time interval, the entire
row will be deleted. This way, reads against the metrics_index table will
always be over live columns and not tombstones.