Add support for server itests on Windows.  Arquillian+Surefire+Cassandra+Windows did not like the way we used the suite deployment extension as a hook for starting and stopping a test cassandra cluster.  The spawned cassandra process seems to confuse things such that the arquillian process and its deployment (eap) process would hang, and the surefire report would not be generated.  Likely due to the fact that Windows does not deal with "background" processes the same way as linux.  This was after already trying to trim the test cluster from 2 nodes to 1 node, for similar process-related issues.
So, on Windows we have decided that to run the i-tests you must use an external
storage node (for example, a dev container instance, or a test product install).
To do this you specify:

    -Ditest.use-external-storage-node=true

By default the external storage node seed is 127.0.0.1|7199|9042.  To specify the
a different storage node upi can specify:

    -Drhq.cassandra.seeds="host|jmxport|cqlport"

This should also work for linux, if desired.

Also:
- Cassandra on Windows will not generate the pid files that were in
place to handle the killing of the test cassandra.  So, I added an enhancement
to the ProcessExecution utilities to return the actual Process, which is now
used to destroy the spawned cassandra processes.  Note that did not end up
being used in the final windows solution.  But I left it as is, because I think
even for linux it's more simple and safer than the pid file approach.

- Added autoimport of "RHQ Storage Node"
- Added test for autoimport of "RHQ Storage Node"
- Also, for windows, I had changed around the way we hooked the storage node
  handling into CCMSuiteDeploymentExtension. Again, not used in the final
  solution but I think it's a bit safer/clearer so I'll leave as is.