changing transaction demarcation for basline and OOB calculations
The queries and methods in MeasurementBaselineManagerBean and in
MeasurementOOBManagerBean were too coarse-grained. In a test environment, I
was seeing regular timeouts while querying for schedules that do not yet have
baselines. And the method that computes OOBs was experiencing frequent
transaction timeouts. The transaction demarcation has been changed so that we
are doing a small, fixed size amount of working for each transaction when
calculating baselines and OOBs. Calls to Cassandra have been moved to
non-transactional methods.

Lastly, fixing a bug in MetricsBaselineCalculator to make sure we always return
a baseline for each schedule that it receives.