Fix test that started failing after the change to remove JOIN FETCH from Criteria calls that include paging and optional data.  It seems the new approach changes slightly the way the data is returned.
loadDefinition() returns hibernate proxied data. It is not safe to pass
proxies into equals() impls because these typically access fields directly,
not through getters, and therefore bypass the proxied data.

Note that this change in behavior due to the JOIN FETCH fix would only
affect non-client code (like test code or intra-slsb calls) because return
values to client code would already be cleansed of the proxies.  The test
is fixed in the same way, by using HibernateDetachUtility to scrub the
proxies before attempting the assertEquals().