Bug 966665 - Criteria queries with paging and sorting on non unique fields may return same item on different pages
All paged criteria queries now apply ID as the least significant
ordering field (see exceptions below).  Since ID is unique, this ensures
that queries for pages always return the same ordering of rows.

ID will not be automatically appended as an ordering field if:
- the paging is unlimited (i.e. no pagesize, all rows returned)
- the Criteria class does not support an ID sort
- the caller has set the Criteria class instance to not support the ID sort
- ID is already specified as an ordering field
- the max number of ordering fields has already been set (3)

Note that CriteriaQuery has had its support for implicit ID sorting removed,
as the new mechanism, which moves support to the CriteriaQueryGenerator,
replaces it.