handle race condition during raw data inserts
MeasurementDataManagerBean passes a callback to MetricsServer.addData(). That
callback is invoked once for each raw data object that is persisted and then
after all of the data has been persisted. The callback populates a list with
the inserted raw data, and that list is then used to update the alert
notification cache. I was seeing NPEs in AgentConditionCache.checkConditions
because the collection passed to it contained null elements. Through some
tedious debugging I beleive that this is a result of concurrent inserts to the
list in the aforementione callback. A lock is now used to serialize inserts in
that list.