[BZ 967542] - UI shows confusing units when editing Availability Duration alert condition type
I improved the DurationItem component to use internally the SelectItem
instead of ComboBoxItem. When using ComboBoxItem, it was possible to
write to the combobox and this is something that shouldn't happen in
this component because the unit types should be fixed. The DurationItem
is also used from the operation tab view, but no reggression was
introduced by this commit in this view.

Modifying also the DurationItem.setAndFormatValue() to populate the time correctly when
!isReadOnly.