set hostname of dev-container storage node to localhost by default
If the hostname option is not set, the storage installer defaults to
InetAddress.getLocalHost().getHostName() which may not return localhost
depending on your hosts configuration. This in turn has led to some problems
with people being able to connect to Cassandra. By explicitly setting the
hostname to 127.0.0.1, I hope we will avoid those issues in the future. Also
note that I explicitly chose 127.0.0.1 instead of localhost since localhost
aliases (e.g., 127.0.0.1, 127.0.0.2, etc.) are used to run multiple nodes on
the same box.