refactor query to find schedules without baselines
The JPQL query previously used resulted in a very inefficient cross join that
could result in statement timeouts in the database with a just a moderately
large number of schedules. This commit removes MeasurementBaselineTest since it
only tested the query that has been replaced. There is already test coverage
of the query in MeasurementBaselineManagerTest.