adding support to update seeds property in addNodeMaintenance resource operation
There is an issue in how the yaml document is serialized. It fails to parse.
I am relying on the Cassandra config classes for reading/writing the yaml
config file; so, we may have to explore an alternative approach. Cassandra
reads but does not write the yaml file, and we are having problems with
how the file is being written.