make sure seed nodes are accounted for and persisted at start up
I have renamed SessionManagerBean to StorageClientManagerBean to better reflect
its purpose. The new init method which is called at server start up does a few
things including,

* making sure storage node connection info is available
Storage node info has to be available in either the rhq.cassandra.seeds property
or in the database in order to initialize the session object. An exception is
thrown if neither exists.

* persist seed nodes if necessary
Seed nodes which are specified in rhq-server.properties will be persisted if
not already in the database. This is neecessary in case we have an HA
deployment where another rhq server does not have the seeds property set in
rhq-server.properties.

* Detect if any new seed nodes have been specified
If there are new need nodes (that is, that do not match what we have in the db)
we treat this as a topology change. This could happen for instance in an HA
deployment where after installing my first server/storage node, I then install
a second server/storage node.