add shutdown operation for storage client subsystem
When the CassandraClusterHeartBeatJob puts the server into maintenance mode it
now invokes StorageClientManagerBean to shut down the storage client subsystem.
When it takes the server out of maintenance mode, it calls
StorageClientManagerBean to reinitialize the storage client subsystem. This is
done in large part to ensure that the cassandra driver session gets recycled so
that we can reconnect to C* when it comes back up.

I have also increased the interal of the heart beat job from 5 sec to 1 minute.
5 seconds was pretty agressive and I probably had it set to that fast of an
interval to faciliate testing.