add support for aggregating data that was collected prior to server start up
This commit addresses scenarios where raw data has been persisted and then the
server is shut down before that data is aggregated. The server could be offline
for 5 minutes, 5 hours, or 5 days. In each of those scenarios, at server start
up we detect that we have missed computing aggregates. And then when the data
purge job runs next, we compute aggregates for 1hr, 6hr, and 24hr data.