only check if the full duration of the time slice has completed once
There is no need to check if the full duration of the time slice has elapsed
for each entry in the index for a give bucket. The check only needs to be
performed once, and there is no need to query metrics_index for the check.
This will substantially reduce the number of queries that are executed during
aggregation.