fix tests that were failing because of the storage client failing to initialize properly
StorageClientManagerBean runs at server start up to bootstrap the storage
client, and it expects a server entity to be in the database as would be the
case with a regular deplomyment. This commit adds functionality to the test
suite such that a Server entity is persisted before any tests are executed.