drop default cassandra user and perform schema changes with rhqadmin user
Previously we were applying RF (replication_factor) changes after creating the
rhqadmin user. This would result in inconsistent reads when there are multiple
nodes and we wind up increasing the RF. The inconsistent reads means that we
would wind up with failed authentication some times. The easiest fix is to
increase the RF before we create the rhqadmin user.