default to the IP address instead of hostname for listen_address property
Cassandra uses the IP address for gossip communication, and during various
maintenance tasks we need to compare addresses with the addresses we have in
StorageNode entities. By defaulting to the hostname as previously did, we have
a mapping layer we have to go through. Using the IP address allows us to avoid
that extra work.