[BZ 917765] - Symlinks in deploy dir no longer confuse relative file path
if deployDir was:
/opt/depls/target -> /opt/realdepls/realtarget
(i.e. target was a symlink to realtarget)

and a file was to be deployed to:
../conf/file

a resulting path would be:
opt/depls/target/../conf/file
which would be resolved as:
opt/realdepls/conf/file
(i.e. the symlink would be resolved first and then the path would be
normalized)

This is not how people usually understand the paths and how they expect
the bundle deployer to function.

We therefore first manually normalize the path and only then obtain its
absolute path to point to it during the deployment process.

Notice that we no longer use the canonical paths anywhere in the deployer
code.