big refactoring of the work that is done when a new storage node is committed to inventory
Here is a run down of the changes.

* fix endpoint comparisons
There was a bug in StorageModeMaintenanceJob.waitForClustering where it was
comparing the string form of an ip address against a PropertySimple.toString.
It should be comparing the value of the property.

* schedule the addNodeMaintenance op as a group operation
StorageNodeMaintenanceJob now schedules the operations for each storage node as
a group operation instead of as individual operations. This allows us to remove
a lot of code around waiting for operations to complete before scheduling the
next one. More importantly, the previous implementation was not blocking until
each operation completed which resulted in repair operations running across
multiple nodes simultenously. We definitely want to run repair on the nodes
serially. Scheduling the work as a group operation handles that for us.

Scheduling the work as a group operation required a slight change to
StorageNodeManagerBean.linkResource. The resource has to be added to the group
before the quartz job is scheduled. Logic needs to be added to verify that the
node has actually joined to cluster before adding it to the group.

* add logic to detect when repair needs to run
There is logic in place now to determine whether or not repair needs to run.
Previously, we would run repair against each node whenever
StorageNodeMaintenanceJob would run. We only want to run repair if and when we
have to since it is a very resource intensive operation.

* Add logic back to update replication_factor of system_auth keyspace
I had previously changed the logic in TopologyManager to *not* update the RF of
the system_auth keyspace. For an multi-node installation, we would increase the
RF of system_auth because the change was made after the rhqadmin user was
created. Without running repair this results in inconsistent reads which in
turns leads to failed authentication. When StorageNodeMaintenanceJob runs we do
want to update the RF of both the system_auth and rhq keysapces. I have
refactored TopologyManager so that system_auth gets updated.