Update the way the percentage disk space use is calculated. It is now based on the aggregate usage of all the partitions that have data files. Also, this is based on the total disk use, not just Cassandra. Added an alert based on this metric.
Added a new metric to the storage node composite that aggregates all the metrics that refer to storage related to Cassandra: data files, commit logs, row cache, and key cache. This will be used for CLI an web only, it will not be a plugin metric.

Updated the plugin descriptor to collect metrics related to disk space usage.

Minor other code refactorings and updates.