remove usage of cassandra-env.sh for cassandra start up
We want to manage C* heap settings and other options/arguments passed to the
C* JVM command line. On linux, the JVM options are set using the JVM_OPTS
shell script variable which is done in cassandra-env.sh. The cassandra-env.sh
script however does not lend itself well to our structured, resource
configuration. The JVM_OPTS variable is now built up in
cassandra-jvm.properties which is a java properties file that also uses legal
shell syntax. The file is sourced from the cassandra start script. This paves
the way for resource configuration of the Cassandra JVM options.

I am also hoping that we can do something similar for Windows where we can use
cassandra-jvm.properties. This would substantially reduce the need for platform
specific code in the storage plugin.