temporarily disable the quartz job that is kicked off when a new node is added
I need to disable the job while working on the internode authentication stuff.
If the maintenance operations run on the nodes, the cluster can actually get
into a bad state because the nodes' authentication conf files have not been
updated which means the new node is not actually part of the cluster.