Bug 886126 - Datasource JNDI change takes too long to be propagated to JON server
In EAP5, JNDI name is the key of Datasource and Connection Factory managed components. Consequently, there is no way to change the JNDI name of such components without creating a new managed component, which will in the end be discovered by RHQ.

So in RHQ, jndi-name config property is now marked as read-only (it will still be required when creating a Datasource or a Connection Factory).