check cluster status using cql driver instead of using jmx
The server had been polling storage cluster nodes with a jmx call to make sure
that at least one node is up for client requests. If no nodes are up, then the
server goes into maintenance mode.

That check is no longer done with a quartz job using jmx. It is now performed
with event notifications we receive from the cql driver. This means that the
server no longer has to care about storage node jmx ports. This has a couple
benefits. First, it reduces the complexity involved with changing the jmx port.
Secondly, since the server no longer is making jmx calls to storage nodes, the
port can be locked down to localhost access.

Any server side code that needs access to the driver's Session should instead
use the new StorageSession class. StorageSession provides the same API, and it
also encapsulates the event handling/propagation logic that is necesary for
monitoring the cluster availability.