add storage node shutdown operation that uses pid file
There can be problems with getting the actual, current pid from Sigar, so the
storage node plugin will first attempt to get the pid from the pid file which
should exist on disk. If that fails, then we will attempt to get the pid from
Sigar.

This implementation for shutdown does not work for Windows so need to revist it
to add in Windows support. We may even consider delegating to rhqtcl.