adding storage_port to plugin config for storage/cassandra plugins
storage_port (i.e., gossip) is not really a connection setting. It has been
added to facilitate initializing shared, cluster settings on the server. We
need to store the cluster settings on the server so that when new storage nodes
are deployed, we know what cluster settings to assign them.

I am working on initializing the cluster settings when the first storage ndoe
is imported into inventory. Prior to this commit, the gossip port was only
available via the storage node's resource configuration, which is not
available when the storage node is imported into inventory. By storing the
cluster settings in the resource's plugin config, the cluster settings can
easily be initialized at the point when the storage node is committed into
inventory.