chema manager is not usable inside the running server
The schema manager classes loads XML files from the classpath. It expects the
URL protocol to be either file or jar. When running inside the server, the
protocol is VFS. I put a TODO in the code to update it and added an else block
to throw an exception for any other protocols that are not supported.
Previously UpdateFolder was basically failing silently and late which made
debugging difficult.