refactor how we determine the most recently stored raw data
I was previuosly using the server creation time to determine how far back to
query the metrics_index table to find the most recent raw data. That logic
breaks in an HA envrionment with multiple servers.

This change simplifies things a bit. We only need to check up until the duration
of the raw retention period which is 7 days. Anything older than that will be
purged.