attempting to fix server itest failures
I am not sure why but DiscoveryBossBeanTest has been failing due to calls to
StorageClusterSettingsManagerBean made by StorageNodeManagerBean. I am
deploying a test impl for StorageClusterSettingsManager to see if it resovles
the issues. There is no need to deploy the production EJB. It is just a thin
wrapper around SystemManagerBean, and it can/should be tested separately.