refactoring state transitions and adding method for deployment
When a storage node is committed into inventory its operation mode is set to
INSTALLED, unless the storage node entity exists in which case the mode is set
to NORMAL. After creating the storage node entity, deployment is started. The
operation mode changs to ANNOUNCE. The address of the new node is announced to
existing cluster nodes. After announcing completes, the operation mode changes
to BOOTSTRAP, and the prepareForBootstrap operation is run on the new node.
When the new node is reported up as part of the cluster, the operation mode of
all cluster modes is set to ADD_NODE_MAINTENANCE. The addNodeMaintenance
operation is then run on each storage node. When that operation completes, the
node's operation mode is set back to NORMAL.

The StorageNodeManagerBean.deployStorageNode method looks at the operation mode
of the node to determine at what step in the process to start the deployment.
The deployStorageNode method is the only method that the UI or remote API will
need to invoke to start or resume a deployment.