add fields in StorageNode for error reporting during maintenance
Storage node deployment and undeployment consists of a series of different
resource operations. The (un)deployment work flow could fail due to one of
those resource operations. When that occurs we can provide a direct link in the
StorageNode.failedOperation field to the operation history of the failed
operation. This direct link will help with providing quick insight into the
cause of the failure.

There is also a new errorMessage field in StorageNode. This field will provide
summary info about the failure.  If the failure is in server side processing
and not in a resource operation, then the errorMessage field should be set but
not the failedOperation field.