[BZ 958166] Remove dependency on com.sun class
We no longer are performing the JMX calls to C* from the RHQ server but we
still are making JMX calls from the Storage installer. As long as the sun.jdk
module exports,

<path name="com/sun/jndi/url/rmi"/>

We do not need to have the dependency on the com.sun classes. And our build
updates the sun.jdk module to export that path.