make sure the mode is updated at the end of each (un)deployment phase
The transaction demarcation has been updated to ensure that upon successful
completion of a (un)deployment phase, the storage node's mode will be updated
before starting the next phase. Previously it was done in the same transaction,
and if starting the next phase failed, then the node's mode would still be set
to the previously completed phase. This did not break anything, but it did
result in extra work.